fix: Helicopter movement
