+++ include/screen.h
@@ -50,8 +50,8 @@
 
 #define MINX           1      // min screen horizontal pos
 #define MINY           1      // min screen vertical pos
-#define MAXX           80     // max screen horizontal pos
-#define MAXY           24     // max screen vertical pos
+#define MAXX           120     // max screen horizontal pos
+#define MAXY           34     // max screen vertical pos
 
 /**
  * Screen Colors type
