+++ src/main.c
@@ -2,49 +2,16 @@
 #include "screen.h"
 #include "keyboard.h"
 #include "timer.h"
+#include "helicopter.h"
 
 
 int x = 0, y = 0;
 
-void printHelicopter(int *x, int *y) {
-    screenSetColor(CYAN, DARKGRAY);
-    screenGotoxy(*x, *y);
-    printf("     ______\n");
-    screenGotoxy(*x, *y + 1);
-    printf(" |X\\_.-╨─｡_\n");
-    screenGotoxy(*x, *y + 2);
-    printf(" ─´‾‾‷\\  °★¯]─\n");
-    screenGotoxy(*x, *y + 3);
-    printf("         ‾‾‾‾\n");
-}
-
-void moveHelicopter(int *command, int *YPos, int *XPos) 
-{
-    screenSetColor(CYAN, DARKGRAY);
-    if (*command == 119 && *YPos > MINY + 5){
-        *YPos -= 1;
-    } else if (*command == 115 && *YPos < MAXY - 5){
-        *YPos += 1;
-    }
-    
-    if (*command == 97 && *XPos > MINX + strlen(" ─´‾‾‷\\  °★¯]─\n")){
-        *XPos -= 1;
-    } else if (*command == 100 && *XPos < MAXX - strlen(" ─´‾‾‷\\  °★¯]─\n") - 1){
-        *XPos += 1;
-    }
-
-    screenDrawBorders();
-    screenGotoxy(*XPos, *YPos);
-    printHelicopter(XPos, YPos);
-    // Atualiza a tela
-    screenUpdate();
-}
-
 // Inicialização da biblioteca na main
 int main() {
     static int ch = 0;
-    int YPos = 10;
-    int XPos = 10;
+    int YPos = 20;
+    int XPos = 20;
     
     screenInit(1);
     keyboardInit();
@@ -58,11 +25,6 @@ int main() {
 
             ch = readch();
             moveHelicopter(&ch, &YPos, &XPos);
-
-            screenClear();
-            screenInit(1);
-            screenGotoxy(((randomNumber()*327)%44), (randomNumber()%30));
-            printf("0");
             // Atualiza a tela
             screenUpdate();
         }
